// src/app/features/animal-health/round-master/models/viewRoundMaster.model.ts
export interface ViewRoundMaster {
    roundNo:          number;
    stateCd:          number;
    stateName:        string;
    diseaseCd:        number;
    diseaseDesc:      string;
    fromDate:         string;
    toDate:           string;
    dataEntryEndDate: string;
    remarks:          string;
}